+++ src/tools/parser/CMakeLists.txt
